fix: play animal damage sounds for all clients
The damage sounds from the `CharacterSoundComponent` are only played for the owner of that entity, i.e., they are intended for sounds that are audible to the player, but not to others.

In contrast, the damage sounds from the `DamageSound` component are played for all clients, not just the owner of the entity.

Therefore, we move the damage sound specification from `CharacterSound` to `DamageSound` in the animal prefabs that define damage sound assets.